+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Es ersetzt die **Release Notes** aus den Anfangstagen von jFachwert und orientie
 genauso wie an [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Patch-Versionen nicht extra aufgeführt, sondern in der Minor-Version.
 
+## [Unreleased]
+### Added
+- of(..)-Methoden in Bankverbindung ergänzt
+
 ## [4.4.2] - 2023-01-09
 ### Fixed
 - Default-Implementierung der verify-Methode (KSimpleValidator) wandelt wieder `javax.validation.ValidationException` in `IllegalArgumentException` um